updated config.sh to create a sym link only when it's not already created

+++ config.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 
-ln -s ~/trench/dotfiles/bin ~/bin
+if [[ ! -h "$HOME/bin" ]]; then
+    ln -s ~/trench/dotfiles/bin ~/bin
+    exit 0
+fi
+
+# sym link zsh config
 rm ~/.zshrc
 rm ~/.zsh_profile
 ln -s ~/trench/dotfiles/zsh/.zshrc ~/.zshrc
 ln -s ~/trench/dotfiles/zsh/.zsh_profile ~/.zsh_profile
-cp -r ~/trench/dotfiles/config/nvim ~/.config/
-cp -r ~/trench/dotfiles/config/tmux ~/.config/
-cp -r ~/trench/dotfiles/config/ghostty ~/.config/
